requirements.txt
dateutils
future
csv
pdfminer
Pillow
pymssql
python-dateutil
SQLAlchemy
git+git://github.com/BarraQDA/sqlalchemy-sqlany
